Decoupled the texture and shader specs away from the graphics manager.
