+++ Core/Rendering/GfxFactory2D.hpp
@@ -67,7 +67,7 @@ namespace Barrage
         - An invalid shader ID if an error occurred.
     */
     /*************************************************************************/
-    GfxManager2D::ShaderID CreateShader(const char* const filepaths[GfxManager2D::ShaderStage::NUM_SHADERS_POSSIBLE]) const;
+    GfxManager2D::ShaderID CreateShader(const char* const filepaths[ShaderStage::NUM_SHADERS_POSSIBLE]) const;
     /*************************************************************************/
     /*!
       \brief
